separate out actions and reducers, formik does not dispatch yet
