+++ src/actions.js
@@ -1,18 +1,6 @@
 import { SubmissionError } from 'redux-form';
 import { REACT_APP_API_BASE_URL } from './config';
 
-export const FETCH_CARDS_SUCCESS = 'FETCH_CARDS_SUCCESS';
-export const fetchCardSuccess = cards => ({
-	type: FETCH_CARDS_SUCCESS,
-	cards
-});
-
-export const FETCH_CARDS_ERROR = 'FETCH_CARDS_ERROR';
-export const fetchCardError = error => ({
-	type: FETCH_CARDS_ERROR,
-	error
-});
-
 export const ADD_CARD_TO_DECK = 'ADD_CARD_TO_DECK';
 export const addCardToDeck = cardId => ({
 	type: ADD_CARD_TO_DECK,
@@ -43,57 +31,11 @@ export const fetchSavedDeckSuccess = deck => ({
 	deck
 });
 
-const BASE_URL = `https://api.magicthegathering.io/v1/cards`;
-export const fetchCardsFromMtgApi = (key, searchTerm) => dispatch => {
-	return fetch(`${BASE_URL}/?${key}=${searchTerm}`, {
-		method: 'GET',
-		headers: { 'Content-Type': 'application/json' }
-	})
-		.then(res => {
-			if (!res.ok) {
-				if (
-					res.headers.has('content-type') &&
-					res.headers.get('content-type').startsWith('application/json')
-				) {
-					console.log(res.json());
-					return res.json().then(err => Promise.reject(err));
-				}
-				return Promise.reject({
-					code: res.status,
-					message: res.statusText
-				});
-			}
-			return res.json();
-		})
-		.then(res => {
-			let newRes = res.cards.map(card => {
-				return {
-					name: card.name,
-					castingcost: card.manaCost,
-					color: card.colors,
-					type: card.type,
-					id: card.id,
-					text: card.text
-				};
-			});
-			dispatch(fetchCardSuccess(newRes));
-		})
-		.catch(err => {
-			const { reason, message, location } = err;
-			if (reason === 'ValidationError') {
-				return Promise.reject(
-					new SubmissionError({
-						[location]: message
-					})
-				);
-			}
-			return Promise.reject(
-				new SubmissionError({
-					_error: 'Error submitting message'
-				})
-			);
-		});
-};
+export const FETCH_CARDS_ERROR = 'FETCH_CARDS_ERROR';
+export const fetchCardError = error => ({
+	type: FETCH_CARDS_ERROR,
+	error
+});
 
 export const saveDeck = (newDeck, key) => dispatch => {
 	return fetch(REACT_APP_API_BASE_URL, {
